+++ src/include/sll/types.h
@@ -260,6 +260,10 @@ typedef void* sll_library_handle_t;
 
 
 
+typedef void* sll_process_handle_t;
+
+
+
 typedef const sll_char_t* sll_json_parser_state_t;